# systems/if_line/docs/api/story-path-authoring.openapi.yaml
openapi: 3.1.0
info:
  title: IF Line StoryPath Authoring API
  version: 1.0.0
  description: >-
    Replacement authoring and publication contract. This contract is mounted
    directly below /api. Legacy chapter-index authoring routes are not part of
    this API and receive no compatibility adapter.
servers:
  - url: /api
security:
  - cookieAuth: []
...
paths:
...
  /story-paths/{story_path_id}/chapters:
    parameters:
      - {$ref: '#/components/parameters/StoryPathId'}
    get:
      tags: [Chapters]
      operationId: listPathChapters
      responses:
        '200':
          description: Chapters ordered for one path.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PathChapter'}
    post:
      tags: [Chapters]
      operationId: appendPathChapter
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PathChapterCreate'}
      responses:
        '201':
          description: New stable path chapter and slot.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PathChapter'}
        '409': {$ref: '#/components/responses/Conflict'}
  /story-paths/{story_path_id}/chapter-generation-batches:
    parameters:
      - {$ref: '#/components/parameters/StoryPathId'}
    post:
      tags: [Chapters]
      operationId: generateChapterBatch
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChapterBatchRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
  /path-chapters/{path_chapter_id}/generations:
    parameters:
      - {$ref: '#/components/parameters/PathChapterId'}
    post:
      tags: [Chapters]
      operationId: generateChapter
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GenerationRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
  /path-chapters/{path_chapter_id}/revisions:
    parameters:
      - {$ref: '#/components/parameters/PathChapterId'}
    get:
      tags: [Chapters]
      operationId: listChapterRevisions
      responses:
        '200':
          description: Revisions in the path chapter's slot.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ChapterRevision'}
    post:
      tags: [Chapters]
      operationId: createChapterRevision
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChapterRevisionCreate'}
      responses:
        '201':
          description: Unselected chapter revision.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChapterRevision'}
  /path-chapters/{path_chapter_id}/head:
    parameters:
      - {$ref: '#/components/parameters/PathChapterId'}
    get:
      tags: [Chapters]
      operationId: getChapterHead
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Chapters]
      operationId: updateChapterHead
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HeadUpdate'}
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '409': {$ref: '#/components/responses/Conflict'}
  /chapter-revisions/{chapter_revision_id}:
    parameters:
      - {$ref: '#/components/parameters/ChapterRevisionId'}
    get:
      tags: [Chapters]
      operationId: getChapterRevision
      responses:
        '200':
          description: Immutable chapter revision.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChapterRevision'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /story-paths/{story_path_id}/checkpoints/{node_id}/candidate-set-generations:
    parameters:
      - {$ref: '#/components/parameters/StoryPathId'}
      - {$ref: '#/components/parameters/NodeId'}
    post:
      tags: [Candidates]
      operationId: generateCandidateSet
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CandidateSetGenerationRequest'}
      responses:
        '202': {$ref: '#/components/responses/TaskAccepted'}
        '409': {$ref: '#/components/responses/Conflict'}
  /story-paths/{story_path_id}/checkpoints/{node_id}/candidate-set-revisions:
    parameters:
      - {$ref: '#/components/parameters/StoryPathId'}
      - {$ref: '#/components/parameters/NodeId'}
    get:
      tags: [Candidates]
      operationId: listCandidateSetRevisions
      responses:
        '200':
          description: Candidate set history at this path checkpoint.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/CandidateSetRevision'}
  /story-paths/{story_path_id}/checkpoints/{node_id}/candidate-set-head:
    parameters:
      - {$ref: '#/components/parameters/StoryPathId'}
      - {$ref: '#/components/parameters/NodeId'}
    get:
      tags: [Candidates]
      operationId: getCandidateSetHead
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [Candidates]
      operationId: updateCandidateSetHead
      parameters:
        - {$ref: '#/components/parameters/IfMatch'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/HeadUpdate'}
      responses:
        '200': {$ref: '#/components/responses/RevisionHead'}
        '409': {$ref: '#/components/responses/Conflict'}
  /candidate-set-revisions/{candidate_set_revision_id}/candidates:
    parameters:
      - {$ref: '#/components/parameters/CandidateSetRevisionId'}
    get:
      tags: [Candidates]
      operationId: listBranchCandidates
      responses:
        '200':
          description: Immutable candidates in one set revision.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/BranchCandidate'}
  /branch-candidates/{candidate_id}/story-paths:
    parameters:
      - {$ref: '#/components/parameters/CandidateId'}
    post:
      tags: [Candidates]
      operationId: promoteBranchCandidate
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StoryPathPromotion'}
      responses:
        '201':
          description: Newly created child path.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StoryPath'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
...
    StoryPathId:
      name: story_path_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    PathChapterId:
      name: path_chapter_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    ChapterRevisionId:
      name: chapter_revision_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    CandidateSetRevisionId:
      name: candidate_set_revision_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    CandidateId:
      name: candidate_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
    NodeId:
      name: node_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/UUID'}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255}
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current integer lock version, optionally quoted.
      schema: {type: string, pattern: '^("[1-9][0-9]*"|[1-9][0-9]*)$'}
  responses:
...
    RevisionHead:
      description: Current reviewed revision pointer.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/RevisionHead'}
    TaskAccepted:
      description: Durable asynchronous task accepted.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/TaskAccepted'}
...
    NotFound:
      description: Resource is absent or not visible to this caller.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: Lock, idempotency, provenance, or source fingerprint conflict.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
    UUID:
      type: string
      format: uuid
...
    GenerationRequest:
      type: object
      properties:
        instructions: {type: [string, 'null'], maxLength: 12000}
        parameters: {type: object, additionalProperties: true}
...
    RevisionHead:
      type: object
      required: [revision_id, lock_version]
      properties:
        revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        lock_version: {type: integer, minimum: 1}
        updated_at: {type: string, format: date-time}
    HeadUpdate:
      type: object
      required: [revision_id]
      properties:
        revision_id: {$ref: '#/components/schemas/UUID'}
...
    StoryPathPromotion:
      type: object
      properties:
        title: {type: [string, 'null'], maxLength: 200}
...
    StoryPath:
      type: object
      required: [id, project_id, title, status, lock_version]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        project_id: {type: integer}
        parent_path_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        fork_path_chapter_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        fork_checkpoint_node_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        fork_candidate_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        base_state_snapshot_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        title: {type: string}
        status: {type: string, enum: [active, archived]}
        lock_version: {type: integer, minimum: 1}
...
    PathChapterCreate:
      type: object
      properties:
        after_path_chapter_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
    PathChapter:
      type: object
      required: [id, story_path_id, chapter_slot_id, display_index, lock_version]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        story_path_id: {$ref: '#/components/schemas/UUID'}
        chapter_slot_id: {$ref: '#/components/schemas/UUID'}
        display_index: {type: integer, minimum: 1}
        predecessor_path_chapter_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        inherited_from_path_chapter_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        current_revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        lock_version: {type: integer, minimum: 1}
    ChapterBatchRequest:
      type: object
      required: [path_chapter_ids]
      properties:
        path_chapter_ids:
          type: array
          minItems: 1
          uniqueItems: true
          items: {$ref: '#/components/schemas/UUID'}
        instructions: {type: [string, 'null'], maxLength: 12000}
    ChapterRevisionCreate:
      type: object
      required: [content]
      properties:
        parent_revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        content: {type: string}
    ChapterRevision:
      type: object
      required: [id, chapter_slot_id, created_for_story_path_id, revision_no, context_manifest, context_hash, content_hash, content]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        chapter_slot_id: {$ref: '#/components/schemas/UUID'}
        created_for_story_path_id: {$ref: '#/components/schemas/UUID'}
        parent_revision_id: {oneOf: [{$ref: '#/components/schemas/UUID'}, {type: 'null'}]}
        revision_no: {type: integer, minimum: 1}
        context_manifest: {type: object, additionalProperties: true}
        context_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
        content: {type: string}
    CandidateSetGenerationRequest:
      type: object
      required: [chapter_revision_id, state_snapshot_id, candidate_count]
      properties:
        chapter_revision_id: {$ref: '#/components/schemas/UUID'}
        state_snapshot_id: {$ref: '#/components/schemas/UUID'}
        candidate_count: {type: integer, minimum: 2, maximum: 4}
        instructions: {type: [string, 'null'], maxLength: 12000}
    CandidateSetRevision:
      type: object
      required: [id, story_path_id, checkpoint_node_id, chapter_revision_id, state_snapshot_id, revision_no, source_hash, content_hash]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        story_path_id: {$ref: '#/components/schemas/UUID'}
        checkpoint_node_id: {$ref: '#/components/schemas/UUID'}
        chapter_revision_id: {$ref: '#/components/schemas/UUID'}
        state_snapshot_id: {$ref: '#/components/schemas/UUID'}
        revision_no: {type: integer, minimum: 1}
        source_hash: {type: string}
        content_hash: {type: string, pattern: '^[0-9a-f]{64}$'}
    BranchCandidate:
      type: object
      required: [id, candidate_set_revision_id, option_key, preview_text, state_delta]
      properties:
        id: {$ref: '#/components/schemas/UUID'}
        candidate_set_revision_id: {$ref: '#/components/schemas/UUID'}
        option_key: {type: string}
        preview_text: {type: string}
        state_delta: {type: object, additionalProperties: true}
...
    TaskAccepted:
      type: object
      required: [task_id, status, events_url, created]
      properties:
        task_id: {$ref: '#/components/schemas/UUID'}
        status: {type: string}
        events_url: {type: string}
        created: {type: boolean}
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true}